+++ src/xml/xmlclient.cpp
@@ -86,7 +86,7 @@ bool XmlClientHandler::GenerateRequest(const char* method, Value& params, Stream
     return true;
 }
 
-ProcessResponseEnum XmlClientHandler::ProcessResponse(char* response, int length, Value& result, unsigned requestId, bool notification)
+ProcessResponseEnum XmlClientHandler::ProcessResponse(char* response, size_t length, Value& result, unsigned requestId, bool notification)
 {
     log_trace();
     ProcessResponseEnum processResponse = ProcessResponseErrorClose;
